create lifecycle/ folder
